+++ kernel/src/Memory/PhysicalPageFrameAllocator.cpp
@@ -128,7 +128,7 @@ namespace WorldOS {
             }
             if (!BitmapAddressFound) {
                 fprintf(VFS_DEBUG, "Bitmap error!\n");
-                Panic("Failed to find address for page bitmap", nullptr, false);
+                PANIC("Failed to find address for page bitmap");
                 return;
             }
         }